+++ client/my-sites/stats/stats-notices/feedback-notice.tsx
@@ -1,5 +1,9 @@
 import NoticeBanner from '@automattic/components/src/notice-banner';
 import { useTranslate } from 'i18n-calypso';
+import { useState } from 'react';
+import useNoticeVisibilityMutation from 'calypso/my-sites/stats/hooks/use-notice-visibility-mutation';
+import useNoticeVisibilityQuery from 'calypso/my-sites/stats/hooks/use-notice-visibility-query';
+import { StatsNoticeProps } from './types';
 
 interface FeedbackNoticeProps {
 	onTakeSurveyClick: () => void;
@@ -7,7 +11,7 @@ interface FeedbackNoticeProps {
 	onNoticeDismiss: () => void;
 }
 
-const FeedbackNotice = ( {
+export const FeedbackNoticeBody = ( {
 	onTakeSurveyClick,
 	onRemindMeLaterClick,
 	onNoticeDismiss,
@@ -42,6 +46,53 @@ const FeedbackNotice = ( {
 			</NoticeBanner>
 		</div>
 	);
+};
+
+const FeedbackNotice = ( { siteId }: StatsNoticeProps ) => {
+	const [ noticeDismissed, setNoticeDismissed ] = useState( false );
+	const { data: showFeedbackNotice } = useNoticeVisibilityQuery( siteId, 'new_stats_feedback' );
+
+	const { mutateAsync: dismissNewStatsFeedbackAsync } = useNoticeVisibilityMutation(
+		siteId,
+		'new_stats_feedback'
+	);
+	const { mutateAsync: dismissFeedbackNoticeAsync } = useNoticeVisibilityMutation(
+		siteId,
+		'new_stats_feedback'
+	);
+	const { mutateAsync: postponeNewStatsFeedbackAsync } = useNoticeVisibilityMutation(
+		siteId,
+		'new_stats_feedback',
+		'postponed',
+		30 * 24 * 3600
+	);
+	const onTakeSurveyClick = () => {
+		dismissNewStatsFeedbackAsync();
+		// Leave some time for the notice to be dismissed.
+		setTimeout(
+			() =>
+				( window.location.href =
+					'https://jetpack.com/redirect?source=new-jetpack-stats-usage-survey' ),
+			250
+		);
+	};
+	const dismissFeedbackNotice = () => {
+		setNoticeDismissed( true );
+		dismissFeedbackNoticeAsync();
+	};
+
+	const onRemindMeLaterClick = () => {
+		setNoticeDismissed( true );
+		postponeNewStatsFeedbackAsync();
+	};
+
+	return ! noticeDismissed && showFeedbackNotice ? (
+		<FeedbackNoticeBody
+			onTakeSurveyClick={ onTakeSurveyClick }
+			onRemindMeLaterClick={ onRemindMeLaterClick }
+			onNoticeDismiss={ dismissFeedbackNotice }
+		/>
+	) : null;
 };
 
 export default FeedbackNotice;
