+++ client/my-sites/stats/stats-notices/types.ts
@@ -1,0 +1,3 @@
+export interface StatsNoticeProps {
+	siteId: number | null;
+}
